Squashed commit which encapsulates looping-queue
Simplifies quite a bit in ShootingGallery. You'll notice how this
simplifies the interface to initializing the spawn queue for balls, and
removes the need for some variables and a pretty ugly method.

Though, it was mostly unnecessary. I just wanted to go on a short
refactoring-adventure. There's really more important things to do
right now.

commit 07a7f08f48779afd88d9d13d102fff919fee7a4b
Author: Wilhelmsen <wwilhelmsen@sogn.no>
Date:   Thu Oct 16 00:27:46 2025 +0200

    Utilize the new LoopingQueue in ShootingGallery

commit 2a0fd1d41257b5477d31be28dd2e1146e7f010c2
Author: Wilhelmsen <wwilhelmsen@sogn.no>
Date:   Thu Oct 16 00:14:02 2025 +0200

    Implement looping queue class

commit d698c94c29f574ff233081e2f48bb2e7e87714ea
Author: Wilhelmsen <wwilhelmsen@sogn.no>
Date:   Wed Oct 15 23:51:52 2025 +0200

    Start classification of looping queue functionality